AFTER STATEMENT + NOTIFY only to "wake up" the flusher
PostgreSQL LISTEN/NOTIFY is probably unreliable at the Python part -- when Python part calls tasks there is probably a stampede of other notifies comming which get lost. This way using a payload with record ID we achieve exactly nothing. NOTIFY is now used only to wake up the task flusher
and the flushing is done in the background (in the Celery queue).